the-wild-cabin

+++ src/services/apiCabins.js
@@ -1,37 +1,50 @@
-import supabase from "./supabase";
-import { supabaseUrl } from "./supabase";
-// API used for creating a cabins
-export const getCabins = async function () {
-  let { data, error } = await supabase.from("cabins").select("*");
+import supabase, { supabaseUrl } from "./supabase";
+
+export async function getCabins() {
+  const { data, error } = await supabase.from("cabins").select("*");
+
   if (error) {
     console.error(error);
-    throw new Error("cabin could not be loaded");
+    throw new Error("Cabins could not be loaded");
   }
 
   return data;
-};
-//API USED FOR CREATING CUBINS
-export const CreateCabins = async function (newCabin) {
+}
+
+async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     "",
   );
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  //create new cabin
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
+
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
+
   if (error) {
     console.error(error);
-    throw new Error("cabin could not be created");
+    throw new Error("Cabin could not be created");
   }
 
-  //upload file
+  // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
+
   // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
@@ -40,13 +53,18 @@ export const CreateCabins = async function (newCabin) {
       "Cabin image could not be uploaded and the cabin was not created",
     );
   }
+
   return data;
-};
-// API used for deleting a cabin
-export const deleteCabin = async function (id) {
-  const { error } = await supabase.from("cabins").delete().eq("id", id);
+}
+
+export async function deleteCabin(id) {
+  const { data, error } = await supabase.from("cabins").delete().eq("id", id);
+
   if (error) {
     console.error(error);
-    throw new Error("cabin could not be deleted");
+    throw new Error("Cabin could not be deleted");
   }
-};
+
+  return data;
+}
+export default createEditCabin;
